+++ samples/project/Kotlin/src/main/kotlin/godot/samples/games/catchBall/BounceOff.kt
@@ -5,9 +5,9 @@ import godot.core.*
 
 class BounceOff: Area() {
 
-    override fun _process(delta: Float) {
+    override fun _process(delta: Double) {
         val arr = getOverlappingBodies()
         if(arr.size() > 0)
-            (RigidBody from (arr[0] ?: throw ArrayIndexOutOfBoundsException())).bounce = 0f
+            (RigidBody from (arr[0] ?: throw ArrayIndexOutOfBoundsException())).bounce = 0.0
     }
 }
