+++ samples/project/Kotlin/src/main/kotlin/godot/samples/games/dodge/Player.kt
@@ -24,7 +24,7 @@ class Player: Area2D() {
         this.connect("hit", getParent(), "gameOver")
     }
 
-    override fun _process(delta: Float) {
+    override fun _process(delta: Double) {
         var velocity = Vector2()
         if (Input.isActionPressed("ui_right"))
             velocity.x += 1
@@ -34,25 +34,25 @@ class Player: Area2D() {
             velocity.y -= 1
         if (Input.isActionPressed("ui_down"))
             velocity.y += 1
-        if (velocity.length() != 0f) {
-            velocity = velocity.normalized() * speed.toFloat()
+        if (velocity.length() != 0.0) {
+            velocity = velocity.normalized() * speed.toDouble()
             playerSprite.play()
         }
         else
             playerSprite.stop()
 
         position += velocity * delta
-        if (position.x < 0) position.x = 0f
+        if (position.x < 0) position.x = 0.0
         else if (position.x > screensize.x) position.x = screensize.x
-        if (position.y < 0) position.y = 0f
+        if (position.y < 0) position.y = 0.0
         else if (position.y > screensize.y) position.y = screensize.y
 
-        if (velocity.x != 0f) {
+        if (velocity.x != 0.0) {
             playerSprite.setAnimation("right")
             playerSprite.flipV = false
             playerSprite.flipH = velocity.x < 0
         }
-        else if (velocity.y != 0f) {
+        else if (velocity.y != 0.0) {
             playerSprite.setAnimation("up")
             playerSprite.flipH = false
             playerSprite.flipV = velocity.y > 0
