+++ samples/project/Kotlin/src/main/kotlin/godot/samples/games/shmup/Bullet.kt
@@ -6,7 +6,7 @@ import godot.core.Vector2
 class Bullet: Area2D() {
     var speed = 1000
 
-    override fun _process(delta: Float) {
+    override fun _process(delta: Double) {
         position -= Vector2(0, speed * delta)
     }
 }
