+++ samples/project/Kotlin/src/main/kotlin/godot/samples/games/shmup/EnemyPath.kt
@@ -11,7 +11,7 @@ class EnemyPath: Path2D() {
     override fun _ready() {
         val follow = getNode(NodePath("PathFollow2D"))
         addChild(tween)
-        tween.interpolateProperty(follow, NodePath("unit_offset"), Variant(0), Variant(1), 6.0f, Tween.TRANS_LINEAR, Tween.EASE_IN_OUT)
+        tween.interpolateProperty(follow, NodePath("unit_offset"), Variant(0), Variant(1), 6.0, Tween.TRANS_LINEAR.toLong(), Tween.EASE_IN_OUT.toLong())
         tween.setRepeat(true)
         tween.start()
     }
